3.6 rozbudowa calc

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -1,13 +1,13 @@
 package com.kodilla.basic_assertion;
 
 public class Calculator {
-    public int sum(int a , int b) {
+    public double sum(double a , double b) {
         return a + b;
     }
-    public int subtract(int a, int b) {
+    public double subtract(double a, double b) {
         return a - b;
     }
-    static double pow(int a , int b) {
+    static double pow(double a , double b) {
         return Math.pow(a , b);
     }
 }
